First try to implement a storage for points
this works as following:
- a field is a QList of tiles
- a tile is a square area with the size of sizeOfTile and an offset
- a track is a store of numPointsInTrack positions/orientations and
  a bounding box

This layout is done like this, as the base datatype of QVector3D is float;
this allows only a precision of 7 digits. So with a lenght of more than
100m, the presicion is only 1mm anymore. OpenGL also works with floats.
As UTM uses coordinates from 100'000 to 899'999m, a float is too unprecise,
local coordinates are needed. A consistent local coordinate system is to be
done, this is a proof of concept.