Added a NMEA-parser, as of now can only read GGA/GNS sentences
To test it, import minimal.json and then nmea-parser.json. Then replace the
connection between the PoseSimulator and the Transverse Mercator Converter
with a connection from NmeaParser to the TMC. Go to the values tab and then
to the strings, select the string "Filename" and click on the button to
select a file. The FileStream reads the file linewise with a defined
frequency, which can be adjusted with the number "Frequency"