Added a block for the calculations of the Ackermann steering
The block calculates in all directions: from either center, left and right
wheel to center, left and right wheel. This helps later to adjust for
sensors, which are directly mounted on the hub of a wheel. For now, it is
used to calculate the wheel angles for the tractor model.