Use LineMesh and LineMeshGeometry of KDEs Atelier [github.com/KDE/atelier]
These classes are modified for the use here; mainly removal of Atelier-
specific code like GCode-loading.

The grid works now as it should. The old code created an entity with all
components like material and geometry. This created a lot of overhead and
didn't work in the case of a change in the parameters of the grid. The new
code creates one entity with a geometry, a transform and a material. In the
geometry is a buffer for all the endpoints for the lines, which get drawed
with the given material (for the color) and transformation.