Implements trigger a local calculation of the Pose
Implements trigger a calculation of a local Pose to generate the position
of their left and right edges. With this, the global planner can calculate
the resulting swath-width to generate its paths. This enables path planning
for multiple implements like front and back mowers with a simple connection
of the correct edges to the planner.