Import/Export of fields, GIS-software compatible
If you save a field, a new GeoJSON is created with two Features: one of the
type Polygon to save the processed boundary and one with the type
MultiPoint to save the recorded points. This file is a standart GeoJSON and
can be opened in pe. QGIS.
The other way round works too: in QGIS you can create or edit a polygon- or
multipoint-layer, export it to GeoJSON and open that file in
QtOpenGuidance.